feat(tulearn): refactor and modularize Tulearn functionality
- Introduced new modules for database interactions, gamification, learner state, and helper functions.
- Created a dedicated `db.ts` for database client management.
- Added `gamification.ts` to handle XP awarding logic.
- Implemented `learner-state.ts` for managing learner state and heart mechanics.
- Added utility functions in `helpers.ts` for common operations.
- Refactored `service.ts` to export functions from new modules, improving code organization and readability.
- Defined types in `types.ts` for better type safety and clarity across the Tulearn module.
